+++ src/index.ts
@@ -5,24 +5,31 @@ import * as query from './query.js'
 import { GoMulticastDNS } from './compat/index.js'
 import type { PeerDiscovery, PeerDiscoveryEvents } from '@libp2p/interface-peer-discovery'
 import type { PeerInfo } from '@libp2p/interface-peer-info'
-import { Components, Initializable } from '@libp2p/components'
 import { symbol } from '@libp2p/interface-peer-discovery'
 import { stringGen } from './utils.js'
+import type { PeerId } from '@libp2p/interface-peer-id'
+import type { AddressManager } from '@libp2p/interface-address-manager'
 
 const log = logger('libp2p:mdns')
 
-export interface MulticastDNSOptions {
+export interface MulticastDNSInit {
   broadcast?: boolean
   interval?: number
   serviceTag?: string
   peerName?: string
   port?: number
+  ip?: string
   compat?: boolean
   compatQueryPeriod?: number
   compatQueryInterval?: number
 }
 
-export class MulticastDNS extends EventEmitter<PeerDiscoveryEvents> implements PeerDiscovery, Initializable {
+export interface MulticastDNSComponents {
+  peerId: PeerId
+  addressManager: AddressManager
+}
+
+class MulticastDNS extends EventEmitter<PeerDiscoveryEvents> implements PeerDiscovery {
   public mdns?: multicastDNS.MulticastDNS
 
   private readonly broadcast: boolean
@@ -30,31 +37,34 @@ export class MulticastDNS extends EventEmitter<PeerDiscoveryEvents> implements P
   private readonly serviceTag: string
   private readonly peerName: string
   private readonly port: number
+  private readonly ip: string
   private _queryInterval: ReturnType<typeof setInterval> | null
   private readonly _goMdns?: GoMulticastDNS
-  private components: Components = new Components()
+  private readonly components: MulticastDNSComponents
 
-  constructor (options: MulticastDNSOptions = {}) {
+  constructor (components: MulticastDNSComponents, init: MulticastDNSInit = {}) {
     super()
 
-    this.broadcast = options.broadcast !== false
-    this.interval = options.interval ?? (1e3 * 10)
-    this.serviceTag = options.serviceTag ?? '_p2p._udp.local'
-    this.peerName = options.peerName ?? stringGen(63)
+    this.broadcast = init.broadcast !== false
+    this.interval = init.interval ?? (1e3 * 10)
+    this.serviceTag = init.serviceTag ?? '_p2p._udp.local'
+    this.ip = init.ip ?? '224.0.0.251'
+    this.peerName = init.peerName ?? stringGen(63)
     // 63 is dns label limit
     if (this.peerName.length >= 64) {
       throw new Error('Peer name should be less than 64 chars long')
     }
-    this.port = options.port ?? 5353
+    this.port = init.port ?? 5353
+    this.components = components
     this._queryInterval = null
     this._onPeer = this._onPeer.bind(this)
     this._onMdnsQuery = this._onMdnsQuery.bind(this)
     this._onMdnsResponse = this._onMdnsResponse.bind(this)
 
-    if (options.compat !== false) {
-      this._goMdns = new GoMulticastDNS({
-        queryPeriod: options.compatQueryPeriod,
-        queryInterval: options.compatQueryInterval
+    if (init.compat !== false) {
+      this._goMdns = new GoMulticastDNS(components, {
+        queryPeriod: init.compatQueryPeriod,
+        queryInterval: init.compatQueryInterval
       })
       this._goMdns.addEventListener('peer', this._onPeer)
     }
@@ -66,12 +76,6 @@ export class MulticastDNS extends EventEmitter<PeerDiscoveryEvents> implements P
 
   get [Symbol.toStringTag] () {
     return '@libp2p/mdns'
-  }
-
-  init (components: Components): void {
-    this.components = components
-
-    this._goMdns?.init(components)
   }
 
   isStarted () {
@@ -88,7 +92,7 @@ export class MulticastDNS extends EventEmitter<PeerDiscoveryEvents> implements P
       return
     }
 
-    this.mdns = multicastDNS({ port: this.port })
+    this.mdns = multicastDNS({ port: this.port, ip: this.ip })
     this.mdns.on('query', this._onMdnsQuery)
     this.mdns.on('response', this._onMdnsResponse)
 
@@ -105,12 +109,12 @@ export class MulticastDNS extends EventEmitter<PeerDiscoveryEvents> implements P
     }
 
     log.trace('received incoming mDNS query')
-    const localPeerId = this.components.getPeerId()
+    const localPeerId = this.components.peerId
     query.gotQuery(
       event,
       this.mdns,
       this.peerName,
-      this.components.getAddressManager().getAddresses().map((ma) => ma.encapsulate('/p2p/' + localPeerId.toString())),
+      this.components.addressManager.getAddresses().map((ma) => ma.encapsulate('/p2p/' + localPeerId.toString())),
       this.serviceTag,
       this.broadcast)
   }
@@ -175,6 +179,10 @@ export class MulticastDNS extends EventEmitter<PeerDiscoveryEvents> implements P
 
     this.mdns = undefined
   }
+}
+
+export function mdns (init: MulticastDNSInit = {}): (components: MulticastDNSComponents) => PeerDiscovery {
+  return (components: MulticastDNSComponents) => new MulticastDNS(components, init)
 }
 
 /* for reference