+++ test/multicast-dns.spec.ts
@@ -5,18 +5,18 @@ import type { Multiaddr } from '@multiformats/multiaddr'
 import { multiaddr } from '@multiformats/multiaddr'
 import { createEd25519PeerId } from '@libp2p/peer-id-factory'
 import pWaitFor from 'p-wait-for'
-import { MulticastDNS } from './../src/index.js'
+import { mdns } from './../src/index.js'
 import type { PeerId } from '@libp2p/interface-peer-id'
 import type { PeerInfo } from '@libp2p/interface-peer-info'
 import { stubInterface } from 'ts-sinon'
 import type { AddressManager } from '@libp2p/interface-address-manager'
-import { Components } from '@libp2p/components'
+import { start, stop } from '@libp2p/interfaces/startable'
 
 function getComponents (peerId: PeerId, multiaddrs: Multiaddr[]) {
   const addressManager = stubInterface<AddressManager>()
   addressManager.getAddresses.returns(multiaddrs)
 
-  return new Components({ peerId, addressManager })
+  return { peerId, addressManager }
 }
 
 describe('MulticastDNS', () => {
@@ -63,21 +63,18 @@ describe('MulticastDNS', () => {
   it('find another peer', async function () {
     this.timeout(40 * 1000)
 
-    const mdnsA = new MulticastDNS({
+    const mdnsA = mdns({
       broadcast: false, // do not talk to ourself
       port: 50001,
       compat: false
-    })
-    mdnsA.init(getComponents(pA, aMultiaddrs))
+    })(getComponents(pA, aMultiaddrs))
 
-    const mdnsB = new MulticastDNS({
+    const mdnsB = mdns({
       port: 50001, // port must be the same
       compat: false
-    })
-    mdnsB.init(getComponents(pB, bMultiaddrs))
+    })(getComponents(pB, bMultiaddrs))
 
-    await mdnsA.start()
-    await mdnsB.start()
+    await start(mdnsA, mdnsB)
 
     const { detail: { id } } = await new Promise<CustomEvent<PeerInfo>>((resolve) => mdnsA.addEventListener('peer', resolve, {
       once: true
@@ -85,32 +82,27 @@ describe('MulticastDNS', () => {
 
     expect(pB.toString()).to.eql(id.toString())
 
-    await Promise.all([mdnsA.stop(), mdnsB.stop()])
+    await stop(mdnsA, mdnsB)
   })
 
   it('announces all multiaddresses', async function () {
     this.timeout(40 * 1000)
 
-    const mdnsA = new MulticastDNS({
+    const mdnsA = mdns({
       broadcast: false, // do not talk to ourself
       port: 50003,
       compat: false
-    })
-    mdnsA.init(getComponents(pA, aMultiaddrs))
-    const mdnsB = new MulticastDNS({
+    })(getComponents(pA, aMultiaddrs))
+    const mdnsB = mdns({
       port: 50003, // port must be the same
       compat: false
-    })
-    mdnsB.init(getComponents(pB, cMultiaddrs))
-    const mdnsD = new MulticastDNS({
+    })(getComponents(pB, cMultiaddrs))
+    const mdnsD = mdns({
       port: 50003, // port must be the same
       compat: false
-    })
-    mdnsD.init(getComponents(pD, dMultiaddrs))
+    })(getComponents(pD, dMultiaddrs))
 
-    await mdnsA.start()
-    await mdnsB.start()
-    await mdnsD.start()
+    await start(mdnsA, mdnsB, mdnsD)
 
     const peers = new Map()
     const expectedPeer = pB.toString()
@@ -123,32 +115,26 @@ describe('MulticastDNS', () => {
 
     expect(peers.get(expectedPeer).multiaddrs.length).to.equal(3)
 
-    await Promise.all([
-      mdnsA.stop(),
-      mdnsB.stop(),
-      mdnsD.stop()
-    ])
+    await stop(mdnsA, mdnsB, mdnsD)
   })
 
   it('doesn\'t emit peers after stop', async function () {
     this.timeout(40 * 1000)
 
-    const mdnsA = new MulticastDNS({
+    const mdnsA = mdns({
       port: 50004, // port must be the same
       compat: false
-    })
-    mdnsA.init(getComponents(pA, aMultiaddrs))
+    })(getComponents(pA, aMultiaddrs))
 
-    const mdnsC = new MulticastDNS({
+    const mdnsC = mdns({
       port: 50004,
       compat: false
-    })
-    mdnsC.init(getComponents(pD, dMultiaddrs))
+    })(getComponents(pD, dMultiaddrs))
 
-    await mdnsA.start()
+    await start(mdnsA)
     await new Promise((resolve) => setTimeout(resolve, 1000))
-    await mdnsA.stop()
-    await mdnsC.start()
+    await stop(mdnsA)
+    await start(mdnsC)
 
     mdnsC.addEventListener('peer', () => {
       throw new Error('Should not receive new peer.')
@@ -157,39 +143,39 @@ describe('MulticastDNS', () => {
     })
 
     await new Promise((resolve) => setTimeout(resolve, 5000))
-    await mdnsC.stop()
+    await stop(mdnsC)
   })
 
   it('should start and stop with go-libp2p-mdns compat', async () => {
-    const mdns = new MulticastDNS({
+    const mdnsA = mdns({
       port: 50004
-    })
-    mdns.init(getComponents(pA, aMultiaddrs))
+    })(getComponents(pA, aMultiaddrs))
 
-    await mdns.start()
-    await mdns.stop()
+    await start(mdnsA)
+    await stop(mdnsA)
   })
 
   it('should not emit undefined peer ids', async () => {
-    const mdns = new MulticastDNS({
+    const mdnsA = mdns({
       port: 50004
-    })
-    mdns.init(getComponents(pA, aMultiaddrs))
-    await mdns.start()
+    })(getComponents(pA, aMultiaddrs))
+    await start(mdnsA)
 
     await new Promise<void>((resolve, reject) => {
-      mdns.addEventListener('peer', (evt) => {
+      mdnsA.addEventListener('peer', (evt) => {
         if (evt.detail == null) {
           reject(new Error('peerData was not set'))
         }
       })
 
-      if (mdns.mdns == null) {
+      // @ts-expect-error not a PeerDiscovery field
+      if (mdnsA.mdns == null) {
         reject(new Error('mdns property was not set'))
         return
       }
 
-      mdns.mdns.on('response', () => {
+      // @ts-expect-error not a PeerDiscovery field
+      mdnsA.mdns.on('response', () => {
         // query.gotResponse is async - we'll bail from that method when
         // comparing the senders PeerId to our own but it'll happen later
         // so allow enough time for the test to have failed if we emit
@@ -200,7 +186,8 @@ describe('MulticastDNS', () => {
       })
 
       // this will cause us to respond to ourselves
-      mdns.mdns.query({
+      // @ts-expect-error not a PeerDiscovery field
+      mdnsA.mdns.query({
         questions: [{
           name: 'localhost',
           type: 'A'
@@ -208,6 +195,33 @@ describe('MulticastDNS', () => {
       })
     })
 
-    await mdns.stop()
+    await stop(mdnsA)
+  })
+
+  it('find another peer with different udp4 address', async function () {
+    this.timeout(40 * 1000)
+
+    const mdnsA = mdns({
+      broadcast: false, // do not talk to ourself
+      port: 50005,
+      ip: '224.0.0.252',
+      compat: false
+    })(getComponents(pA, aMultiaddrs))
+
+    const mdnsB = mdns({
+      port: 50005, // port must be the same
+      ip: '224.0.0.252', // ip must be the same
+      compat: false
+    })(getComponents(pB, bMultiaddrs))
+
+    await start(mdnsA, mdnsB)
+
+    const { detail: { id } } = await new Promise<CustomEvent<PeerInfo>>((resolve) => mdnsA.addEventListener('peer', resolve, {
+      once: true
+    }))
+
+    expect(pB.toString()).to.eql(id.toString())
+
+    await stop(mdnsA, mdnsB)
   })
 })